Week4/Lanchain-Agent-Task/requirements.txt
# LangChain Core
langchain>=0.1.0
langchain-community>=0.0.13
langchain-aws>=0.1.0

# AWS Bedrock for LLM
boto3>=1.34.0

# MCP
google-api-python-client
google-auth
google-auth-oauthlib
python-docx>=0.8.11
PyPDF2>=3.0.0

# Vector Store for RAG
chromadb>=0.4.22
sentence-transformers>=2.2.2

# Web Search
ddgs>=4.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
